Move Continuation to Core
