+++ modules/core/arrow-core/src/main/kotlin/arrow/core/Continuation.kt
@@ -1,0 +1,11 @@
+package arrow.core
+
+interface Continuation<in T> : kotlin.coroutines.Continuation<T> {
+  fun resume(value: T)
+
+  fun resumeWithException(exception: Throwable)
+
+  override fun resumeWith(result: SuccessOrFailure<T>) =
+    result.fold(::resume, ::resumeWithException)
+
+}
