+++ modules/core/arrow-typeclasses/src/main/kotlin/arrow/typeclasses/Comonad.kt
@@ -1,6 +1,7 @@
 package arrow.typeclasses
 
 import arrow.Kind
+import arrow.core.Continuation
 import arrow.core.identity
 import java.io.Serializable
 import kotlin.coroutines.*
